requirements.txt
aiohappyeyeballs==2.4.3
aiohttp==3.10.10
aiosignal==1.3.1
asgiref==3.8.1
attrs==24.2.0
certifi==2024.8.30
channels==4.2.0
charset-normalizer==3.4.0
dateparser==1.2.0
Django==5.1.2
django-cors-headers==4.6.0
djangorestframework==3.15.2
drf-nested-routers==0.94.1
frozenlist==1.5.0
idna==3.10
multidict==6.1.0
propcache==0.2.0
pybit==5.9.0
pycryptodome==3.21.0
python-binance==1.0.22
python-dateutil==2.9.0.post0
pytz==2024.2
regex==2024.9.11
requests==2.32.3
six==1.16.0
sqlparse==0.5.1
tzlocal==5.2
ujson==5.10.0
urllib3==2.2.3
websocket-client==1.8.0
websockets==13.1
yarl==1.17.1

#  MariaDB connector
mysqlclient==2.1.1

django-audit-log-middleware>=0.0.4
argon2-cffi>=23.1.0
cryptography>=41.0.0
python-dotenv>=1.0.0
djangorestframework-simplejwt==5.3.1
